AsynThunk for new game creation
